backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# Cache
redis==5.0.4

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3

# ML core
scikit-learn==1.5.0
xgboost==2.0.3
pandas==2.2.2
numpy==1.26.4
faiss-cpu==1.8.0
joblib==1.4.2
imbalanced-learn==0.12.3

# Drift detection
scipy==1.13.1

# Data
openpyxl==3.1.5

# Utilities
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
structlog==24.2.0
tenacity==8.3.0
httpx==0.27.0
aiofiles==23.2.1
